fix: Diet 관련 데이터 타입 수정
- DailyDiet의 total_calorie 타입을 long → double로 변경하였습니다.
- DietSetFood의 칼로리 및 영양소 필드를 Double → primitive double로 변경하여 불필요한 박싱을 제거했습니다.

+++ src/main/java/com/project/growfit/domain/Diet/entity/DailyDiet.java
@@ -36,7 +36,7 @@ public class DailyDiet extends BaseEntity {
 
     @Column(nullable = false)
     @JoinColumn(name = "total_calorie")
-    private long totalCalorie;
+    private double totalCalorie;
 
     @OneToMany(mappedBy = "dailyDiet", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Diet> diets = new ArrayList<>();
